Move vote_generator and vote_spacing in to their own files.
Recompile time for voting.hpp before and vote headers after
Before:
[117/117] Linking CXX executable rpc_test
ninja  1262.92s user 87.90s system 869% cpu 2:35.28 total
After:
[16/16] Linking CXX executable nano_node
ninja  98.05s user 11.24s system 377% cpu 28.917 total
