+++ nano/node/active_transactions.hpp
@@ -5,7 +5,6 @@
 #include <nano/node/election_insertion_result.hpp>
 #include <nano/node/election_status.hpp>
 #include <nano/node/vote_with_weight_info.hpp>
-#include <nano/node/voting.hpp>
 #include <nano/secure/common.hpp>
 
 #include <boost/multi_index/hashed_index.hpp>
@@ -17,6 +16,7 @@
 #include <condition_variable>
 #include <deque>
 #include <memory>
+#include <thread>
 #include <unordered_map>
 
 namespace mi = boost::multi_index;
@@ -32,7 +32,14 @@ class election;
 class vote;
 class confirmation_height_processor;
 class stats;
+}
+namespace nano::store
+{
+class read_transaction;
+}
 
+namespace nano
+{
 class recently_confirmed_cache final
 {
 public:
